Refactored methods to work with MongoDB via Mongoose.
